+++ src/utilities/index.ts
@@ -1,4 +1,3 @@
 export {eventPoint} from './event-point';
-export {wrapLabel} from './wrap-label';
 export {getTextWidth} from './get-text-width';
 export {clamp} from './clamp';